requirements.txt
# Core Airflow & Framework (Match your image version)
apache-airflow==2.9.3
apache-airflow-providers-common-sql
apache-airflow-providers-ftp
apache-airflow-providers-http
apache-airflow-providers-imap
apache-airflow-providers-sqlite

# Azure Data Lake and Storage
adlfs
azure-storage-blob
azure-datalake-store
azure-identity
azure-core
azure-synapse-artifacts

# Key Data Processing Libraries
pandas
numpy
sqlalchemy

# HTTP/API Clients
requests
aiohttp
httpx

# Utilities
python-dateutil
python-slugify

# Jupyter Notebook Support
ipython
ipykernel
bs4

# Postgres Support (Critical for your setup)
psycopg2-binary
snowflake-sqlalchemy

# Optional: Uncomment if needed for your DAGs
# fsspec~=2023.9.0
# geopy~=2.4.0
# graphviz~=0.20.0